Instaladores y ejecutables actualizados. Versión 4.11

+++ Linux/InstaladorMicroDNA.sh
@@ -1,8 +1,13 @@
 #!/bin/bash
 # --- Unidad de Microarreglos DNA, IFC, UNAM, MX ---
-#Revisa paquete Zenity
+#Verifica sólo una instancia
+if test -f ~/Downloads/instDNA.txt; then
+    echo .
+    read -t 5 -p "MicroarreglosDNA ya se está instalando, esta ventana se cerrará automáticamente..."
+fi
+#Revisa paquete Zenity y actualiza paquetes
+echo "instDNA">~/Downloads/instDNA.txt
 sudo apt-get install zenity -y
-#Actualiza sistema con paquetes
 zenity --info --title="Instalador Microarreglos DNA" --width=400 --text="Gracias por utilizar 'Microarreglos DNA'. Para evitar brechas de seguridad, se actualizan las dependencias del sistema operativo. \n\n Este proceso tarda alrededor de 40 minutos, aunque suele variar acorde a la velocidad de su internet y cantidad de dependencias por actualizar. \n\n Presione aceptar para continuar."
 sudo apt-get update && sudo apt-get upgrade -y
 sudo apt-get autoremove -y
@@ -16,16 +21,18 @@ zenity --info --title="Microarreglos DNA" --width=400 --text="Para continuar con
 sudo apt-get install apache2 -y
 sudo usermod -a -G www-data pi
 zenity --info --title="Microarreglos DNA" --width=400 --text="Paso 1/3 completo. Servidor local instalado y verificado. \n\n Presione aceptar para continuar."
-#Instala git para el programa
+#Instala git para el programa y verifica programa
 sudo apt-get install git -y
-sudo mkdir /var/www/html/MicroDNA
-cd /var/www/html/MicroDNA/
-sudo git clone "https://github.com/frankfsa6/MicroarreglosDNA.git" 
-sudo mv MicroarreglosDNA ../
-cd ..
-sudo rm -r MicroDNA
+if test ! -d /var/www/html/MicroarreglosDNA; then
+    sudo mkdir /var/www/html/MicroDNA
+    cd /var/www/html/MicroDNA/
+    sudo git clone "https://github.com/frankfsa6/MicroarreglosDNA.git" 
+    sudo mv MicroarreglosDNA ../
+    cd ..
+    sudo rm -r MicroDNA
+fi
 #Define atributos a carpetas 
-sudo chmod -R -f 0707 /var/www/html/MicroarreglosDNA
+sudo chmod -R -f 0777 /var/www/html/MicroarreglosDNA
 sudo chown -R pi:pi /var/www/html/MicroarreglosDNA
 zenity --info --title="Microarreglos DNA" --width=400 --text="Paso 2/3 completo. Gestor de versiones 'Github' y programa 'Microarreglos DNA' instalado correctamente. \n\n Presione aceptar para continuar."
 #Verifica o instala PHP y MySQL/MariaDB
@@ -37,7 +44,19 @@ sudo apt-get install pigpio -y
 sudo apt-get autoremove -y
 sudo mysql < /var/www/html/MicroarreglosDNA/sql/"usuario.sql"
 sudo service apache2 restart
-#Otorga permiso superusuario y manda mensaje exitoso
+#Otorga permiso superusuario y pone última versión
 sudo echo 'www-data ALL=(ALL)NOPASSWD:ALL' | sudo EDITOR='tee -a' visudo
-zenity --info --title="Microarreglos DNA" --width=400 --text="Paso 3/3 completo. Lenguajes programadores y configuraciones extras instaladas y verificadas. \n\n Ha finalizado de instalar el programa. Acuda al 'Manual de usuario' para mayor detalle sobre su uso. Microarreglos DNA se abrirá en seguida."
-sudo exec /var/www/html/MicroarreglosDNA/Linux/MicroarreglosDNA.sh
+cd /var/www/html/MicroarreglosDNA
+sudo git reset --hard origin/master
+sudo git pull
+#Crea accesos directos
+sudo mkdir ~/Desktop/MicroarreglosDNA
+sudo cp -u /var/www/html/MicroarreglosDNA/Linux/MicroarreglosDNA.sh ~/Desktop/MicroarreglosDNA/
+sudo cp -u /var/www/html/MicroarreglosDNA/Linux/*.url ~/Desktop/MicroarreglosDNA/
+sudo chmod -R -f 0777 ~/Desktop/MicroarreglosDNA
+#Abre página de internet y finaliza programa
+sudo service apache2 restart
+zenity --info --title="Microarreglos DNA" --width=400 --text="Paso 3/3 completo. Lenguajes programadores y configuraciones extras instaladas y verificadas. \n\n Ha finalizado de instalar el programa. En su escritorio se instaló el archivo ejecutable para su próximo uso. Acuda al 'Manual de usuario' para mayor detalle sobre su uso. \n\n Microarreglos DNA se abrirá en seguida."
+DISPLAY=:0 chromium-browser "http://localhost/MicroarreglosDNA/"
+xdg-open "http://localhost/MicroarreglosDNA/"
+sudo rm -r ~/Downloads/instDNA.txt
